+++ NEWS.md
@@ -5,6 +5,9 @@
 - Generics changes:
   - `^` will now invoke `tf.square()` or `tf.sqrt()` directly when appropriate
   - `|`, `&`, and `!` now cast arguments to 'bool' dtype.
+  - `print()` now shows 1d shapes without a trailing commas.
+
+
 
 - Tensors with dtype 'string' now convert to R character vectors by methods
   `as.array()` and `as.matrix()`. (previously they converted to python.builtin.bytes,